Pre-review cleanup: fireRemoveMutationAndLeave + getProductNoun + dead code
