+++ client/dashboard/utils/purchase.ts
@@ -800,3 +800,38 @@ export const hasMarketplaceProduct = ( productsList: Product[], searchSlug: stri
 			// SaaS products are also considered marketplace products
 			( product_type.startsWith( 'marketplace' ) || product_type === 'saas_plugin' )
 	);
+
+/**
+ * Plain-English noun for a purchase ("plan" / "domain" / "email" / "theme" /
+ * "plugin" / "subscription"). Used by the cancelled success notice and by the
+ * remove-success snackbar.
+ */
+export function getProductNoun( purchase: Purchase ): string {
+	if ( purchase.is_plan ) {
+		return __( 'plan' );
+	}
+	if ( purchase.is_domain_registration ) {
+		return __( 'domain' );
+	}
+	const slug = purchase.product_slug ?? '';
+	if (
+		slug.startsWith( 'wp_titan_mail' ) ||
+		slug === 'gapps' ||
+		slug === 'gapps_extra_license' ||
+		slug === 'gapps_business' ||
+		slug === 'gapps_business_extra_license' ||
+		slug.startsWith( 'wp_google_workspace_' )
+	) {
+		return __( 'email' );
+	}
+	if ( purchase.product_type === 'marketplace_theme' ) {
+		return __( 'theme' );
+	}
+	if (
+		purchase.product_type?.startsWith( 'marketplace' ) ||
+		purchase.product_type === 'saas_plugin'
+	) {
+		return __( 'plugin' );
+	}
+	return __( 'subscription' );
+}
